Much faster mining, plenty of testing, discord bot

+++ DemCoinCommons/DemCoinSettings.cs
@@ -1,7 +1,8 @@
 namespace DemCoinCommons;
 
 public class DemCoinSettings {
-    public const int Difficulty = 23;  // Number of leading zeroes required in hash  was 26, prob 30 for adam, 23 for testing
+    // 23 is very fast, 32 is reasonable for prod
+    public const int Difficulty = 23;  // Number of leading zeroes (bits) required in hash
     public const double MinerReward = 1;  // Coins rewarded for mining one block
-    public static int AverageHashesToBlock() => (int)Math.Pow(2, Difficulty);  // Average amount of tries before mining one block
+    public static double AverageHashesToBlock() => Math.Pow(2, Difficulty);  // Average amount of tries before mining one block
 }
